merge, resolve conflicts, modify file structure
