+++ README.md
@@ -1,3 +1,13 @@
 # Hackathon Project
 
 team member: Jacky, Jiayue, Sam
+
+# config map API
+/client/.env.local
+REACT_APP_MAP_API_KEY = APIKEY
+
+# how to run 
+/client
+npm start
+/server
+npm run dev
