+++ client/src/components/reusableComponent/searchBar/SearchBar.js
@@ -3,14 +3,17 @@ import { useState } from "react";
 import locationIcon from "../../../assets/images/location.svg";
 import searchIcon from "../../../assets/images/search.svg";
 import micIcon from "../../../assets/images/mic.svg";
+import { useNavigate } from "react-router-dom";
 
 const SearchBar = () => {
+  const navigate = useNavigate();
   return (
     <div className="search-bar">
       <input
         className="search-bar-input"
         type="text"
         placeholder="Search for anything..."
+        onKeyDown={(e) => {e.key === 'Enter' && navigate('/map')}}
       />
       <img src={searchIcon} alt="Search Icon" className="search-icon" />
       <img src={micIcon} alt="Mic Icon" className="mic-icon" />
