+++ client/src/components/tempComponents/apiComponent/Api.js
@@ -1,5 +1,6 @@
-import Nav from '../navComponent/Nav'
+import Nav from '../../reusableComponent/navComponent/Nav'
 
+export default function Api() {
     return (
         <div className="Api">
             <h1>Api</h1>
