+++ client/src/store.js
@@ -1,7 +1,7 @@
 // “redux”包，只完成redux本身的功能
 import { createStore, applyMiddleware, combineReducers } from "redux";
 // 导入自己创建的reducer
-import demoReducers from "./components/demoComponent/demoReducers";
+import demoReducers from "./components/tempComponents/demoComponent/demoReducers";
 import thunk from "redux-thunk";
 
 const rootReducer = combineReducers({
