+++ server/routes/index.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const webRobotRouter = require('./webRobotRouter');
+
+router.use('/', webRobotRouter);
+
+module.exports = router;
